Merge entries with the same key.
Kobo's e-reader software does not guarantee that all matching
entries are displayed, so by merging all entries with the same
key into a single entry, we force them to all be displayed.

This was particularly a problem where sometimes conjugated entries
wouldn't be displayed simply because JMDict also had a stand-alone
entry for that conjugation (without the verb definitions).